Add comprehensive tests for dig_t parsing and serialization formats
- Introduced test_mult_inv_optimized.cpp to validate optimized mult_inv() function.
- Created test_parse_ct_4formats.cpp to verify parsing of four formats: d[N]BM, d#N#BM, dig#N#BM, and dig[N]BM.
- Added test_parse_formats.cpp to ensure all parsing formats work correctly and equivalently.
- Implemented test_parse_impl_ct.cpp to demonstrate compile-time parsing without exceptions using std::expected.
- Developed test_to_cstr.cpp to test the to_cstr() function for compile-time serialization of dig_t.

+++ include/core/internal/auxiliary_functions.hpp
@@ -832,7 +832,7 @@ namespace NumRepr {
     if (n <= 0) {
       return -1; // Not in domain of the function log_base(n)
                  // Domain(log_base) = ]0, +infinity[ = [1, +infinity[
-    } else if (n < base) {
+    } else if (static_cast<std::uint64_t>(n) < base) {
       return 0;
     } else {
       return 1 + int_log(base, n / base);
